Provide a direct link to test credentials in DjNRO
This adds a custom variable _EDITURL that provides a direct link to the
test credentials in DjNRO. It can then be used in the Nagios/Naemon
config to e.g. set a notes_url that looks something like this:

  notes_url  https://eduroam.ac.za/manage/$_SERVICEEDITURI$

If you do this, you should probably set _EDITURL in the template to
"instrealmsmon" so if it's unset in the specific service, the default is
something sane and generic.