Provide a way to opt-out from notifications
The update-* scripts automate the process of setting up monitoring
notifications for various services, based on the content of databases.
However, these databases don't necessarily provide a mechanism for
people to opt-out of these notifications, and we don't want people to
remove contact details. Thus it's important to provide a mechanism for
people to opt-out.