Allow the mtime to be reset, triggering a reload
The update-* scripts use mtime to determine whether changes have been
made and whether the config should be regenerated. We want to allow the
opt-out script to trigger those changes, hence resetting mtime. I've
used the epoch value because that makes it obvious that it's not a
"real" mtime, but rather one that's been artificially manipulated.